rechner

+++ Rechner/src/at/simon/rechner/Rechner.java
@@ -1,20 +1,20 @@
 package at.simon.rechner;
 
 public class Rechner {
-    public double addition(int num1, int num2){
-        int number = num1 + num2;
+    public double addition(double num1, double num2){
+        double number = num1 + num2;
         return number;
     }
-    public double subtraction(int num1, int num2){
-        int number = num1 - num2;
+    public double subtraction(double num1, double num2){
+        double number = num1 - num2;
         return number;
     }
-    public double multiplication(int num1, int num2){
-        int number = num1 * num2;
+    public double multiplication(double num1, double num2){
+        double number = num1 * num2;
         return number;
     }
-    public double division(int num1, int num2){
-        int number = num1 / num2;
+    public double division(double num1, double num2){
+        double number = num1 / num2;
         return number;
     }
 }
